Added template directory
Added a template directory in editor, so now you can add templates to that and they will be added to a project when the plugin is activated. Also added a .gdignore so it no longer shows up in the editor.
I've also added a new template based on my personal implementation of using the yield function.
I've also added the template_generator directory to the gitignore file so it won't show up.
Note : I messed up while doing editing the templates and ended up editing the generated template, so I copied and pasted it into the plugin template, which is why it's saying I've rewritten the whole thing.