Autoload and templates (#48)
* Added template and autoload on plugin activation (#28)

* Push errors and warnings to editor in debug

* Added template and autoload on plugin activation

The plugin now will add the "Ink Template" template to the project the plugin is activated in. It will also create the "__InkRuntime" autoload.
I have done some small tests, I would like to do some more but I haven't used GUT before, I need to familiarise myself with it first.

* Added template directory

Added a template directory in editor, so now you can add templates to that and they will be added to a project when the plugin is activated. Also added a .gdignore so it no longer shows up in the editor.
I've also added a new template based on my personal implementation of using the yield function.
I've also added the template_generator directory to the gitignore file so it won't show up.
Note : I messed up while doing editing the templates and ended up editing the generated template, so I copied and pasted it into the plugin template, which is why it's saying I've rewritten the whole thing.

* Delete script_templates directory

Now it's added to .gitignore, remove it from the GitHub repository as it will be generated from the start.

* Delete ink_template.gd

Co-authored-by: Frédéric Maquin <fred@ephread.com>

* Clean up old PR work

* Use %TS% in templates and ignore docs/ while importing

* Update .gitattributes

Co-authored-by: GreenClovers <35625742+greencloversguy@users.noreply.github.com>